VideoCore: Add static_cast around expressions where the compiler doesn’t deduce the right type.

+++ src/video_core/color.h
@@ -124,7 +124,7 @@ inline u32 DecodeD24(const u8* bytes) {
  * @return Resulting values stored as a Math::Vec2
  */
 inline const Math::Vec2<u32> DecodeD24S8(const u8* bytes) {
-    return { (bytes[2] << 16) | (bytes[1] << 8) | bytes[0], bytes[3] };
+    return { static_cast<u32>((bytes[2] << 16) | (bytes[1] << 8) | bytes[0]), bytes[3] };
 }
 
 /**
